requirements.txt
youtube-transcript-api
langchain
langchain-community
langchain-gemini
faiss-cpu
tiktoken
python-dotenv
sentence-transformers
google-generativeai
gradio
